fixes to interlocked and monitor

+++ CoreLib/System/Threading/Interlocked.cs
@@ -22,7 +22,7 @@ namespace System.Threading
         public static extern int Increment(ref int location);
 
         [MethodImpl(MethodImplOptions.InternalCall)]
-        public static extern int Increment(ref long location);
+        public static extern long Increment(ref long location);
         /******************************
          * Decrement
          *   Implemented: int
@@ -55,7 +55,7 @@ namespace System.Threading
         public static extern object Exchange(ref object location1, object value);
 
         [MethodImpl(MethodImplOptions.InternalCall)]
-        public static extern int Exchange(ref IntPtr location1, IntPtr value);
+        public static extern IntPtr Exchange(ref IntPtr location1, IntPtr value);
 
         [MethodImpl(MethodImplOptions.InternalCall)]
         public static extern T Exchange<T>(ref T location1, T value) where T : class;
